electron_microscopy_tools/requirements.txt
# Core dependencies
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0

# File format handling
mrcfile>=1.3.0  # For MRC format
pyem>=0.5.0    # For EM format
h5py>=3.6.0    # For HDF5 format
zarr>=2.11.0   # For Zarr format
pillow>=9.0.0  # For TIFF/TIF handling
hyperspy>=1.7.0  # For DM3/DM4 formats

# Metadata handling
pyyaml>=6.0
jsonschema>=4.4.0
xmltodict>=0.13.0

# Machine learning for discovery
scikit-learn>=1.0.0
transformers>=4.18.0

# Utilities
tqdm>=4.62.0
click>=8.0.0  # For CLI interface
